root store context and setup splitted
