+++ App.tsx
@@ -5,11 +5,9 @@ import useColorScheme from "./app/utils/useColorScheme";
 import Navigation from "./app/navigation";
 import LoadingScreenComponent from "./app/components/LoadingScreen";
 import { initializeI18n } from "./app/i18n/i18n";
-import {
-  RootStoreProvider,
-  RootStoreInstance,
-  setupRootStore,
-} from "./app/models/root-store/root-store";
+import { RootStoreInstance } from "./app/models/root-store/root-store";
+import { RootStoreProvider } from "./app/models/root-store/root-store-context";
+import { setupRootStore } from "./app/models/root-store/setup-root-store";
 
 export default function App() {
   const colorScheme = useColorScheme();
