+++ app/screens/Main.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { View } from "react-native";
 import { observer } from "mobx-react-lite";
 
-import { useStore } from "../models/root-store/root-store";
+import { useStore } from "../models/root-store/root-store-context";
 import MainComponent from "../components/Main";
 
 export const MainScreen = observer(function Main() {
